add (hacky) speedup for UpdateImage() to not wait for an empty queue (saves 14ms).
this is a quick hack to massively improve perf when updating images, which includes creating new images,
drawing text which has not previously been cached and updating the image data itself.

the previous impl would wait for the queue to be empty, which took ~14ms, which often resulted in
dropped frames.

this new impl caches the data and frees it when the command has been processed. it does this by polling
a fence (which i think is some kind of condivar?).

the result is 14ms -> 0ms (haven't measured ns, but that would vary on the image size).